1、修改UI

+++ DagolfLa/球队新版本/view/球队活动/JGHTotalPriceCell.h
@@ -14,6 +14,6 @@
 @property (weak, nonatomic) IBOutlet UILabel *totalPrice;
 
 
-- (void)configTotalPrice:(NSInteger)total;
+- (void)configTotalPrice:(float)total;
 
 @end
